add Hash Test Cases

+++ src/main/test/com/qinjiangbo/HashTest.java
@@ -50,8 +50,8 @@ public class HashTest {
         System.out.println(Hashing.goodFastHash(1));
         System.out.println(Hashing.murmur3_32(100));
         System.out.println(Hashing.sipHash24(10, 1000));
-        long goodHash1 = Hashing.goodFastHash(10).newHasher().putInt(10).hash().asLong();
-        long goodHash2 = Hashing.goodFastHash(10).newHasher().putInt(11).hash().asLong();
+        int goodHash1 = Hashing.goodFastHash(10).newHasher().putInt(10).hash().asInt();
+        int goodHash2 = Hashing.goodFastHash(10).newHasher().putInt(11).hash().asInt();
         System.out.println("goodHash1: " + Hashing.consistentHash(goodHash1, 2));
         System.out.println("goodHash2: " + Hashing.consistentHash(goodHash2, 2));
     }
